+++ src/main/java/com/umc5th/muffler/domain/expense/controller/HomeController.java
@@ -1,7 +1,7 @@
-package com.umc5th.muffler.domain.home.controller;
+package com.umc5th.muffler.domain.expense.controller;
 
-import com.umc5th.muffler.domain.home.dto.WholeCalendarResponse;
-import com.umc5th.muffler.domain.home.service.HomeService;
+import com.umc5th.muffler.domain.expense.dto.homeDto.WholeCalendarResponse;
+import com.umc5th.muffler.domain.expense.service.HomeService;
 import com.umc5th.muffler.global.response.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
