+++ src/main/java/com/umc5th/muffler/domain/expense/dto/homeDto/CategoryCalendarDailyInfo.java
@@ -1,4 +1,4 @@
-package com.umc5th.muffler.domain.home.dto;
+package com.umc5th.muffler.domain.expense.dto.homeDto;
 
 import com.umc5th.muffler.entity.constant.Level;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class DailyCategoryInfoDto {
+public class CategoryCalendarDailyInfo {
 
     private Long dailyTotalCost;
     private Level dailyRate;
