+++ src/main/java/com/umc5th/muffler/domain/expense/dto/homeDto/WholeCalendarDailyInfo.java
@@ -1,4 +1,4 @@
-package com.umc5th.muffler.domain.home.dto;
+package com.umc5th.muffler.domain.expense.dto.homeDto;
 
 import com.umc5th.muffler.entity.constant.Level;
 import lombok.Builder;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Builder
 @Getter
-public class DailyInfoDto {
+public class WholeCalendarDailyInfo {
 
     private Long dailyBudget;
     private Long dailyTotalCost;
